almost finalized pa03
assignments 1, 2, 3, 5 done.

In 4, Sigma_hat_y is asked for in the description of the task, but lütkepohl pages 96 ff tell us, that we (probably??) need sigma_hat_y_hat which is somewhat more complicated (see page 99). So far, i haven't found a way to implement it.

Also, double check with built-in python function still needs to be done + potential bug investigation.